Reorganize extensions, separate blocks extension from plugin extensions
