+++ libs/block-editor/src/lib/ngx-tiptap.module.ts
@@ -9,7 +9,8 @@ import { SuggestionsComponent } from './suggestions/suggestions.component';
 import { MenuModule } from 'primeng/menu';
 import { CommonModule } from '@angular/common';
 import { CardModule } from 'primeng/card';
-import { ContentletBlockComponent } from './extentions/contentlet-block/contentlet-block.component';
+import { ContentletBlockComponent } from './extentions/blocks/contentlet-block/contentlet-block.component';
+
 import { SuggestionsService } from './services/suggestions.service';
 import { SuggestionListComponent } from './suggestion-list/suggestion-list.component';
 import { ActionButtonComponent } from './extentions/action-button/action-button.component';
