+++ libs/block-editor/src/public-api.ts
@@ -11,4 +11,4 @@ export * from './lib/ngx-tiptap.module';
 
 export * from './lib/AngularRenderer';
 export * from './lib/NodeViewRenderer';
-export * from './lib/extentions/contentlet-block.extension'
+export * from './lib/extentions/blocks/contentlet-block/contentlet-block.extension'
